+++ src/Integration/SystemTools/PdfGeneratorIntegration.php
@@ -79,4 +79,9 @@ class PdfGeneratorIntegration implements PlatformSkillInterface
     {
         return false;
     }
+
+    public function getSetupInstructions(): ?string
+    {
+        return null;
+    }
 }