+++ src/Integration/SystemTools/ReadFileIntegration.php
@@ -67,4 +67,9 @@ class ReadFileIntegration implements PlatformSkillInterface
     {
         return false;
     }
+
+    public function getSetupInstructions(): ?string
+    {
+        return null;
+    }
 }