+++ src/Integration/UserIntegrations/HubSpotIntegration.php
@@ -662,4 +662,9 @@ class HubSpotIntegration implements PersonalizedSkillInterface
     {
         return true;
     }
+
+    public function getSetupInstructions(): ?string
+    {
+        return null;
+    }
 }