+++ src/Integration/UserIntegrations/TrelloIntegration.php
@@ -385,4 +385,9 @@ class TrelloIntegration implements PersonalizedSkillInterface
     {
         return false;
     }
+
+    public function getSetupInstructions(): ?string
+    {
+        return null;
+    }
 }